Renamed WP munging scripts for greater cromulence
The scripts have been renamed for greater clarity
and ease-of-tab-completion. For example,
wp_pageviews_extract.rb is now extract_pageviews.rb.
In general, the idea is verb_noun.extension.

This A) makes sure that no two scripts in a
directory have the same leading characters and
B) makes it easier to understand at a glance what
a script does.